requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
astor==0.8.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cycler==0.11.0
decorator==5.1.1
duckdb==0.8.1
exceptiongroup==1.1.3
executing==1.2.0
fonttools==4.42.1
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.36
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
numpy==1.26.0
openai==0.27.10
packaging==23.1
pandas==1.5.3
pandasai==1.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
protobuf==4.24.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
rich==13.5.3
rpds-py==0.10.3
scipy==1.11.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.20
stack-data==0.6.2
streamlit==1.26.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.6
yarl==1.9.2
zipp==3.16.2
